feature update code login

+++ Services/UserService.cs
@@ -154,7 +154,10 @@ namespace FOLYFOOD.Services
                 };
                 
             }
-            await uplloadFile.DeleteFile(userUpdate.Avartar);
+            if(DbContext.Accounts.FirstOrDefault(x=>x.AccountId == id).Avartar != "https://media.istockphoto.com/id/1300845620/vector/user-icon-flat-isolated-on-white-background-user-symbol-vector-illustration.jpg?s=612x612&w=0&k=20&c=yBeyba0hUkh14_jgv1OKqIH0CCSWU_4ckRkAoy2p73o=")
+            {
+                await uplloadFile.DeleteFile(userUpdate.Avartar);
+            }
             string link = await uplloadFile.UploadFile(file);
             userUpdate.Avartar = link;
             DbContext.Accounts.Update(userUpdate);
